templates/template-eval-harness/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eval-harness"
version = "0.1.0"
description = "ML model evaluation harness"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy==1.26.3",
    "pandas==2.1.4",
    "scikit-learn==1.4.0",
    "pyyaml==6.0.1",
    "jinja2==3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "ruff==0.1.13",
]

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["eval_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
